requirements.txt
fastapi
uvicorn[standard]
openai
langchain
langchain-openai
chromadb
pypdf
pymupdf
python-dotenv
python-multipart
tiktoken
streamlit
pandas
plotly
numpy
